% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{logit_CAVI}
\alias{logit_CAVI}
\title{Variational inference for Bayesian logistic regression using CAVI algorithm}
\usage{
logit_CAVI(X, y, prior, tol = 1e-16, maxiter = 10000, verbose = FALSE)
}
\arguments{
\item{X}{The input design matrix. Note the intercept column vector is assumed included.}

\item{y}{The binary response.}

\item{prior}{Prior for the logistic parameters.}

\item{tol}{The ELBO difference tolerance for conversion.}

\item{maxiter}{The maximum iterations if convergence is not achieved.}

\item{verbose}{A diagnostics flag added by Buckley et al.}
}
\value{
A list containing:
\itemize{
\item error - An error message if convergence failed or the number of iterations to achieve convergence.
\item mu - A vector of posterior means.
\item Sigma - A vector of posterior variances.
\item Convergence - A vector of the ELBO at each iteration.
\item LBDifference - A vector of ELBO differences between each iteration.
\item xi - A vector of log-odds per X row.
}
}
\description{
Variational inference for Bayesian logistic regression using CAVI algorithm
}
