% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{logit_NR}
\alias{logit_NR}
\title{Maximum likelihood for the logistic regression using Newton-Raphson}
\usage{
logit_NR(X, y, tol = 1e-16, beta_start = NULL, maxiter = 10000)
}
\arguments{
\item{X}{The input design matrix. Note the intercept column vector is assumed included.}

\item{y}{The binary response.}

\item{tol}{The Newton-Raphson difference tolerance for conversion.}

\item{beta_start}{The starting guess for beta.}

\item{maxiter}{The maximum number of iterations if convergence is not achieved.}
}
\value{
A list containing:
\itemize{
\item beta - A vector of posterior means.
\item Convergence - A vector of iterations up to convergence.
\item Loglikelihood - A vector of log likelihoods for all iterations.
}
}
\description{
Maximum likelihood for the logistic regression using Newton-Raphson
}
