% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{logit_SVI}
\alias{logit_SVI}
\title{Variational inference for Bayesian logistic regression using SVI algorithm}
\usage{
logit_SVI(X, y, prior, iter, tau, kappa)
}
\arguments{
\item{X}{The input design matrix. Note the intercept column vector is assumed included.}

\item{y}{The binary response.}

\item{prior}{Prior for the logistic parameters.}

\item{iter}{The maximum number of iterations if convergence is not achieved.}

\item{tau}{The delay down-weighting early iterations if >= 0.}

\item{kappa}{The forgetting rate.  Must be between 0.5 and 1.0}
}
\value{
A list containing:
\itemize{
\item mu - A vector of posterior means.
\item Sigma - A vector of posterior variances.
}
}
\description{
Variational inference for Bayesian logistic regression using SVI algorithm
}
