% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vb_pheno.R
\name{plot.VBpheno}
\alias{plot.VBpheno}
\title{Variational Bayes Patient Phenotyping.}
\usage{
\method{plot}{VBpheno}(x, ...)
}
\arguments{
\item{x}{An object of class 'VBpheno'.}

\item{...}{Additional arguments passed to the plot method}
}
\description{
Plot Model Summaries, Biomarker shifts and Model Diagnostics for Variational Bayes Patient Phenotyping.
}
