% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vb_pheno.R
\name{print.vbPheno}
\alias{print.vbPheno}
\title{Variational Bayes Patient Phenotyping.}
\usage{
\method{print}{vbPheno}(x, ...)
}
\arguments{
\item{x}{An object of class 'vbPheno'}

\item{...}{Additional arguments to be passed onto lower-level functions at a later stage of development.}
}
\description{
Print a vbPheno object
}
