% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vb_gmm_cavi.R
\name{VB_GMM_ELBO}
\alias{VB_GMM_ELBO}
\title{Calculate the Evidence Lower Bound (ELBO)}
\usage{
VB_GMM_ELBO(X, p, n, q_post, prior)
}
\arguments{
\item{X}{n x p data matrix (or data frame that will be converted to a matrix).}

\item{p}{number of parameters}

\item{n}{number of rows}

\item{q_post}{A list of the fitted posterior Q family at iteration t.}

\item{prior}{Prior for the GMM parameters.}
}
\value{
elbo the Evidence Lower Bound at iteration t following E and M steps.
}
\description{
Calculate the Evidence Lower Bound (ELBO)
}
