% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vb_gmm_cavi.R
\name{VB_GMM_Expectation}
\alias{VB_GMM_Expectation}
\title{Variational Bayes Expectation step}
\usage{
VB_GMM_Expectation(X, n, q_post)
}
\arguments{
\item{X}{n x p data matrix (or data frame that will be converted to a matrix).}

\item{n}{number of rows in X.}

\item{q_post}{A list of the fitted posterior Q family at iteration t-1.}
}
\value{
A list of the fitted posterior Q family after the E step at iteration t.
}
\description{
Variational Bayes Expectation step
}
