% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vb_gmm_cavi.R
\name{VB_GMM_Init}
\alias{VB_GMM_Init}
\title{Initialise the variational parameters and the hyper parameters}
\usage{
VB_GMM_Init(X, k, n, prior, init, initParams)
}
\arguments{
\item{X}{n x p data matrix (or data frame that will be converted to a matrix).}

\item{k}{guess for the number of mixture components.}

\item{n}{number of rows in X.}

\item{prior}{Prior for the GMM parameters.}

\item{init}{initialize the clusters c("random", "kmeans", "dbscan")}

\item{initParams}{initialization parameters for dbscan.  NULL if dbscan not selected for init.}
}
\value{
A list of the initially fitted posterior Q family
}
\description{
Initialise the variational parameters and the hyper parameters
}
