% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vb_gmm_cavi.R
\name{VB_GMM_Maximisation}
\alias{VB_GMM_Maximisation}
\title{Variational Bayes Maximisation step}
\usage{
VB_GMM_Maximisation(X, q_post, prior)
}
\arguments{
\item{X}{n x p data matrix (or data frame that will be converted to a matrix).}

\item{q_post}{A list of the fitted posterior Q family at iteration t-1.}

\item{prior}{Prior for the GMM parameters.}
}
\value{
A list of the fitted posterior Q family after the M step at iteration t.
}
\description{
Variational Bayes Maximisation step
}
