% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vb_pheno.R
\name{plot.vbPheno}
\alias{plot.vbPheno}
\title{Variational Bayes Patient Phenotyping.}
\usage{
\method{plot}{vbPheno}(x, ...)
}
\arguments{
\item{x}{An object of class 'vbPheno'.}

\item{...}{Additional arguments passed to the plot method}
}
\description{
Plot Model Summaries, Biomarker shifts and Model Diagnostics for Variational Bayes Patient Phenotyping.
}
