% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vb_pheno.R
\name{print.VBpheno}
\alias{print.VBpheno}
\title{Variational Bayes Patient Phenotyping.}
\usage{
\method{print}{VBpheno}(x, ...)
}
\arguments{
\item{x}{An object of class 'VBpheno'}

\item{...}{Additional arguments to be passed onto lower-level functions at a later stage of development.}
}
\description{
Print a VBpheno object
}
