% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vb_pheno.R
\name{runModel}
\alias{runModel}
\title{Run the Variational Bayes patient Phenotyping model}
\usage{
runModel(
  ehr_data,
  gmm_X,
  gmm_k,
  gmm_delta = 1e-06,
  gmm_init = "kmeans",
  gmm_initParams = NULL,
  gmm_maxiters = 200,
  gmm_prior = NULL,
  gmm_stopIfELBOReverse = FALSE,
  gmm_verbose = FALSE,
  logit_X,
  logit_prior,
  logit_tol = 1e-16,
  logit_maxiter = 10000,
  logit_verbose = FALSE
)
}
\arguments{
\item{ehr_data}{The EHR data.}

\item{gmm_X}{The input design matrix. Note the intercept column vector is assumed included.}

\item{gmm_k}{The maximum guess for how many disease components exist.  Minimum is 2 (binary response).}

\item{gmm_delta}{Change in ELBO that triggers algorithm stopping.}

\item{gmm_init}{Initialize the clusters c("random", "kmeans", proportion R{ > 0 and < 1}).}

\item{gmm_initParams}{Parameters for an initialiser requiring its own parameters e.g. dbscan.}

\item{gmm_maxiters}{The maximum iterations for VB GMM.}

\item{gmm_prior}{An informative prior for the GMM}

\item{gmm_stopIfELBOReverse}{Stop the VB iterations if the ELBO reverses direction.}

\item{gmm_verbose}{Print out information per iteration to track progress in case of long-running experiments.}

\item{logit_X}{The input design matrix. Note the intercept column vector is assumed included.}

\item{logit_prior}{An informative prior for the logit}

\item{logit_tol}{Change in ELBO that triggers algorithm stopping.}

\item{logit_maxiter}{The maximum iterations for VB logit.}

\item{logit_verbose}{Print out information per iteration to track progress in case of long-running experiments.}
}
\value{
A list containing:
\itemize{
\item prob - The probability of phenotype given the data and priors.
\item biomarker_shift - A data frame containing the biomarker shifts from normal for the phenotype.
\item sensitivity - The sensitivity and specificity for the dichotomous clinical codes.
\item gmm - The VB GMM results. For details see help(vb_gmm_cavi).
\item logit - The VB Logit results.  For details see help(logit_CAVI).
}
}
\description{
Run the Variational Bayes patient Phenotyping model
}
