% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vb_pheno.R
\name{summary.VBpheno}
\alias{summary.VBpheno}
\title{Variational Bayes Patient Phenotyping.}
\usage{
\method{summary}{VBpheno}(object, ...)
}
\arguments{
\item{object}{An object of class 'VBpheno'.}

\item{...}{Additional arguments to be passed onto lower-level functions at a later stage of development.}
}
\description{
Summary method for class 'VBpheno'
}
